requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorlog==6.7.0
Flask==2.2.2
Flask-JWT-Extended==4.4.3
flask-marshmallow==0.14.0
Flask-SQLAlchemy==2.5.1
greenlet==2.0.1
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
marshmallow==3.17.0
marshmallow-sqlalchemy==0.28.1
packaging==21.3
psycopg2==2.9.3
PyJWT==2.6.0
pyparsing==3.0.9
python-dotenv==0.21.0
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.39
urllib3==1.26.12
Werkzeug==2.2.2
